requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
httpx>=0.27.0
python-dotenv>=1.0.0
